% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matViewPipeline.R
\name{extraAction}
\alias{extraAction}
\title{Process the actions required for special handling as instructed for a case from the set available on the MIMIC code repository
This function creates a materialized view from those available on the MIMIC code repository.  See details about where these come from.}
\usage{
extraAction(x, conn)
}
\arguments{
\item{x}{an XML node}

\item{conn}{your postgres DBI object to connect to the postgres db}
}
\value{
TBD
}
\description{
Process the actions required for special handling as instructed for a case from the set available on the MIMIC code repository
This function creates a materialized view from those available on the MIMIC code repository.  See details about where these come from.
}
\details{
TBD
}
\author{
Jesse D. Raffa
}
