% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matViewPipeline.R
\name{get_special_case}
\alias{get_special_case}
\title{Process special materialized views from the set available on the MIMIC code repository
This function creates a materialized view from those available on the MIMIC code repository.  See details about where these come from.}
\usage{
get_special_case(xml, dbSource, conn)
}
\arguments{
\item{xml}{a character string URL to the location of a XML file containing special handling instructions.}

\item{dbSource}{your dplyr database source (Currently must be postgres).  Defaults to pg_src in the global environment.}

\item{conn}{your postgres DBI object to connect to the postgres db}
}
\value{
a remote processed view (dplyr).
}
\description{
Process special materialized views from the set available on the MIMIC code repository
This function creates a materialized view from those available on the MIMIC code repository.  See details about where these come from.
}
\details{
TBD
}
\author{
Jesse D. Raffa
}
