% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matViewPipeline.R
\name{get_views}
\alias{get_views}
\title{Create *ALL* materialized views from the set available on the MIMIC code repository
This function creates a materialized view from those available on the MIMIC code repository.  See details about where these come from.}
\usage{
get_views(URLlist = "london2016_datathon_cirr/mat_view_urls", con, dplyrDB,
  special = FALSE, specialURL = "")
}
\arguments{
\item{URLlist}{a character string URL to the location of a plaintext file containing a list of all SQL files to create the views}

\item{con}{your postgres DBI object to connect to the postgres db}

\item{dplyrDB}{your dplyr database source (Currently must be postgres).}

\item{special}{binary indicated whether to do the special case materialized views which require some additional care.  Defaults to fall}

\item{specialURL}{a character string URL to the locate of a XML file containing the special handling instructions.}
}
\value{
a list containing all the remote processed views.
}
\description{
Create *ALL* materialized views from the set available on the MIMIC code repository
This function creates a materialized view from those available on the MIMIC code repository.  See details about where these come from.
}
\details{
TBD
}
\examples{
## Setup Connection
# m <- dbDriver("PostgreSQL")
# passwd <- "secret";
# con <- dbConnect(m, user="user", password=passwd, dbname="mimic",host="localhost",port=5674) # The trickiest part
# dbSendStatement(con,"SET search_path to mimiciii"); # or whatever schema you have
# pg_src <- src_postgres(dbname = "mimic", host = "127.0.0.1", port = 5674, user = "user", password = passwd,options="-c search_path=mimiciii")
# myViews <- get_views("https://raw.githubusercontent.com/MIT-LCP/mimic-code/master/concepts/code-status.sql",dbSource=pg_src,conn=con)

}
\author{
Jesse D. Raffa
}
